REFACTOR: Decompose Large Chat Components
**Current:** Three components with too many responsibilities (410, 420, 320 lines).
**Future:** Logic extracted to hooks and sub-components.

18. Extract `useStreamResponse()` from chat-ui.tsx — 200-line streaming event loop + abort + tool tracking (*parallel with 19, 20*)
19. Extract `<ToolCallDisplay />`, `<ThinkingDisplay />`, `<MessageActions />` from message-bubble.tsx (*parallel with 18, 20*)
20. Extract `usePromptCommands()` from chat-input.tsx — `/` trigger, filtering, keyboard nav (~80 lines) (*parallel with 18, 19*)
