+++ components/chat/chat-ui.tsx
@@ -3,7 +3,6 @@
 import { useAppStore } from "@/lib/store";
 import type { Message } from "@/types/message";
 import type { Attachment } from "@/types/attachment";
-import { uploadAttachment } from "@/lib/actions/attachments";
 import { reconstructThread, getDeepestLeaf } from "@/lib/chat/tree-utils";
 import { useCallback, useEffect, useRef, useState } from "react";
 import { MessageBubble } from "./message-bubble";
@@ -11,8 +10,7 @@ import { ChatInput } from "./chat-input";
 import { SideView } from "./side-view";
 import { ScrollArea } from "@/components/ui/scroll-area";
 import { v4 as uuidv4 } from "uuid";
-import { persistMessage } from "@/lib/actions/chats/persist-message";
-import { toast } from "sonner";
+import { useStreamResponse } from "@/hooks/chat/use-stream-response";
 import { DEFAULT_MODEL } from "@/models";
 import { Bot } from "lucide-react";
 
@@ -59,24 +57,26 @@ export function ChatUI({
     : undefined;
 
   const [sideViewContent, setSideViewContent] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState(false);
-  const [streamingContent, setStreamingContent] = useState<string | null>(null);
-  const [streamingReasoning, setStreamingReasoning] = useState<string | null>(
-    null,
-  );
-  const [isStreamingReasoning, setIsStreamingReasoning] = useState(false);
-  const [activeToolCalls, setActiveToolCalls] = useState<
-    Array<{
-      toolCallId: string;
-      toolName: string;
-      args: unknown;
-      result?: unknown;
-      status: "calling" | "complete";
-    }>
-  >([]);
+
+  const {
+    isLoading,
+    streamingContent,
+    streamingReasoning,
+    isStreamingReasoning,
+    activeToolCalls,
+    streamResponse,
+    stopStream,
+  } = useStreamResponse(chatId, {
+    onDone: (content) => {
+      const mermaidMatch = content.match(/```mermaid\n([\s\S]*?)```/);
+      if (mermaidMatch) {
+        setSideViewContent(mermaidMatch[0]);
+      }
+    },
+  });
+
   const scrollRef = useRef<HTMLDivElement>(null);
   const sentInitial = useRef(false);
-  const abortControllerRef = useRef<AbortController | null>(null);
 
   useEffect(() => {
     if (mcpServers.length === 0) {
@@ -117,16 +117,6 @@ export function ChatUI({
     scrollToBottom,
   ]);
 
-  const handleStop = useCallback(() => {
-    abortControllerRef.current?.abort();
-  }, []);
-
-  useEffect(() => {
-    return () => {
-      abortControllerRef.current?.abort();
-    };
-  }, []);
-
   if (!chat)
     return (
       <div className="flex h-full items-center justify-center">
@@ -138,253 +128,6 @@ export function ChatUI({
     ? reconstructThread(chat.messages, chat.currentLeafId)
     : [];
 
-  const streamResponse = async (
-    userMsgId: string,
-    content: string,
-    parentId: string | null,
-    attachments: Attachment[] = [],
-    model = DEFAULT_MODEL,
-    selectedServerIds: string[] = [],
-    selectedPromptId?: string,
-  ) => {
-    abortControllerRef.current?.abort();
-    const controller = new AbortController();
-    abortControllerRef.current = controller;
-
-    let fullContent = content;
-    let userMsgMetadata: string | null = null;
-    if (selectedPromptId) {
-      const prompts = useAppStore.getState().prompts;
-      const selectedPrompt = prompts.find((p) => p.id === selectedPromptId);
-      if (selectedPrompt) {
-        fullContent = selectedPrompt.content + "\n\n" + content;
-        userMsgMetadata = JSON.stringify({
-          promptId: selectedPromptId,
-          userContent: content,
-        });
-      }
-    }
-
-    addMessage(
-      chatId,
-      "user",
-      fullContent,
-      parentId,
-      userMsgId,
-      userMsgMetadata,
-      attachments,
-    );
-
-    try {
-      await persistMessage(chatId, {
-        id: userMsgId,
-        role: "user",
-        content: fullContent,
-        parentId,
-        metadata: userMsgMetadata ?? undefined,
-      });
-    } catch (err) {
-      console.error("Failed to persist message:", err);
-      toast.error(
-        "Message may not have been saved. Please check your connection.",
-      );
-    }
-
-    // Upload attachments to server
-    const uploadedAttachments: Attachment[] = [];
-    for (const att of attachments) {
-      try {
-        const formData = new FormData();
-        // Convert dataUrl back to File for upload
-        const response = await fetch(att.dataUrl);
-        const blob = await response.blob();
-        const file = new File([blob], att.name, { type: att.mimeType });
-        formData.append("file", file);
-        formData.append("messageId", userMsgId);
-
-        const data = await uploadAttachment(formData);
-        uploadedAttachments.push({ ...att, id: data.id, key: data.key });
-      } catch {
-        // Upload failed silently — attachment was already shown locally
-      }
-    }
-
-    if (uploadedAttachments.length > 0) {
-      updateMessageAttachments(chatId, userMsgId, uploadedAttachments);
-    }
-
-    const latestChat = useAppStore.getState().chats[chatId];
-    const latestThread = latestChat?.currentLeafId
-      ? reconstructThread(latestChat.messages, latestChat.currentLeafId)
-      : [];
-    const history = latestThread.map((m) => ({
-      role: m.role,
-      content: m.content,
-      attachments: m.attachments?.map((att) => ({
-        id: att.id,
-        type: att.type,
-        dataUrl: att.dataUrl,
-        name: att.name,
-        mimeType: att.mimeType,
-        extractedText: att.extractedText,
-        key: att.key,
-      })),
-    }));
-
-    let accumulated = "";
-    let accumulatedReasoning = "";
-
-    try {
-      const response = await fetch("/api/chat", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({
-          chatId,
-          userMessageId: userMsgId,
-          messages: history,
-          model,
-          selectedServerIds,
-        }),
-        signal: controller.signal,
-      });
-
-      if (!response.ok || !response.body) {
-        if (response.status === 429) {
-          toast.error("Too many requests. Please try again later.");
-          throw new Error("Rate limit exceeded");
-        }
-        if (response.status === 401) {
-          toast.error("Your session has expired. Please log in again.");
-          throw new Error("Unauthorized");
-        }
-        const errorData = await response.json().catch(() => ({}));
-        throw new Error(errorData.error || "Stream request failed");
-      }
-
-      const reader = response.body.getReader();
-      const decoder = new TextDecoder();
-      let buffer = "";
-
-      while (true) {
-        const { done, value } = await reader.read();
-        if (done) break;
-
-        buffer += decoder.decode(value, { stream: true });
-        const lines = buffer.split("\n");
-        buffer = lines.pop() ?? "";
-
-        for (const line of lines) {
-          const trimmed = line.trim();
-          if (!trimmed || !trimmed.startsWith("data: ")) continue;
-          const event = JSON.parse(trimmed.slice(6)) as {
-            type: string;
-            delta?: string;
-            id?: string;
-            message?: string;
-            code?: string;
-            toolCallId?: string;
-            toolName?: string;
-            args?: unknown;
-            result?: unknown;
-            metadata?: { toolCalls: unknown[]; toolResults: unknown[] };
-          };
-
-          if (event.type === "reasoning" && event.delta) {
-            accumulatedReasoning += event.delta;
-            setStreamingReasoning(accumulatedReasoning);
-            setIsStreamingReasoning(true);
-          } else if (event.type === "text" && event.delta) {
-            setIsStreamingReasoning(false);
-            accumulated += event.delta;
-            setStreamingContent(accumulated);
-          } else if (
-            event.type === "tool-call" &&
-            event.toolCallId &&
-            event.toolName
-          ) {
-            setIsStreamingReasoning(false);
-            setActiveToolCalls((prev) => [
-              ...prev,
-              {
-                toolCallId: event.toolCallId!,
-                toolName: event.toolName!,
-                args: event.args,
-                status: "calling",
-              },
-            ]);
-          } else if (event.type === "tool-result" && event.toolCallId) {
-            setActiveToolCalls((prev) =>
-              prev.map((tc) =>
-                tc.toolCallId === event.toolCallId
-                  ? { ...tc, status: "complete" as const, result: event.result }
-                  : tc,
-              ),
-            );
-          } else if (event.type === "done" && event.id) {
-            const metadata = event.metadata
-              ? JSON.stringify(event.metadata)
-              : null;
-            addMessage(
-              chatId,
-              "assistant",
-              accumulated,
-              userMsgId,
-              event.id,
-              metadata,
-              undefined,
-              accumulatedReasoning || undefined,
-            );
-            const mermaidMatch = accumulated.match(/```mermaid\n([\s\S]*?)```/);
-            if (mermaidMatch) {
-              setSideViewContent(mermaidMatch[0]);
-            }
-            setStreamingContent(null);
-            setStreamingReasoning(null);
-            setIsStreamingReasoning(false);
-            setActiveToolCalls([]);
-          } else if (event.type === "error") {
-            setStreamingContent(null);
-            setActiveToolCalls([]);
-            toast.error(event.message || "An error occurred during generation");
-            addMessage(
-              chatId,
-              "assistant",
-              event.message ||
-                "Sorry, I couldn't generate a response. Please try again.",
-              userMsgId,
-            );
-          }
-        }
-      }
-    } catch (err: any) {
-      if (err.name === "AbortError") {
-        toast.info("Generation stopped");
-        if (accumulated.trim()) {
-          addMessage(
-            chatId,
-            "assistant",
-            accumulated,
-            userMsgId,
-            undefined,
-            null,
-            undefined,
-            accumulatedReasoning || undefined,
-          );
-        }
-      } else {
-        console.error("Stream error:", err);
-        toast.error(err.message || "Failed to generate response");
-      }
-    } finally {
-      abortControllerRef.current = null;
-      setStreamingContent(null);
-      setStreamingReasoning(null);
-      setIsStreamingReasoning(false);
-      setActiveToolCalls([]);
-      setIsLoading(false);
-    }
-  };
-
   const handleSend = async (
     content: string,
     attachments: Attachment[] = [],
@@ -392,7 +135,6 @@ export function ChatUI({
     selectedServerIds: string[] = [],
     selectedPromptId?: string,
   ) => {
-    setIsLoading(true);
     await streamResponse(
       uuidv4(),
       content,
@@ -416,7 +158,6 @@ export function ChatUI({
   const handleEdit = async (id: string, newContent: string) => {
     const msg = chat.messages[id];
     if (!msg) return;
-    setIsLoading(true);
     await streamResponse(uuidv4(), newContent, msg.parentId, []);
   };
 
@@ -530,7 +271,7 @@ export function ChatUI({
             <ChatInput
               onSend={handleSend}
               isLoading={isLoading}
-              onStop={handleStop}
+              onStop={stopStream}
               servers={mcpServers.filter((s) => s.enabled)}
             />
           </div>
